+++ utils/react-test/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@ma-shop/is": "0.0.11",
     "enzyme": "3.11.0",
+    "enzyme-adapter-react-16": "1.15.2",
     "prettier": "2.0.5"
   }
 }
