+++ utils/react-test/src/index.mjs
@@ -2,4 +2,5 @@
 // !!! This file is generated do not modify it manually because it will be overwritten
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
+export * from './enzyme'
 export * from './react-test'
